[CEP-21] Move schema management to TCM
Schema itself is a component of ClusterMetadata and all DDL updates are
applied by the CMS inserting a log entry containing a schema
transformation. As the log entries are disseminated around the cluster,
each peer applies the transformation to its local ClusterMetadata,
enacting the schema change. This entails some changes to the way the
database objects represented in schema are intialised (db objects refers
to classes like Keyspace, ColumnFamilyStore, etc).

Co-authored-by: Marcus Eriksson <marcuse@apache.org>
Co-authored-by: Alex Petrov <oleksandr.petrov@gmail.com>
Co-authored-by: Sam Tunnicliffe <samt@apache.org>