[CEP-21] Restore operation mode reporting (7/7)
Part 7 of 7 brings StorageService.operationMode back into sync with
previous behaviour. Many external coordination tools depend on accessing
this state via JMX, so this is an important external interface.

This commit also adds a virtual version of the system.local table, as we
can fully construct the data for this from ClusterMetadata, meaning we no
longer the on-disk system table, though this is retained for now. In
future, more system tables can be virtualised (system.peers,
system_schema, etc).

Co-authored-by: Marcus Eriksson <marcuse@apache.org>
Co-authored-by: Alex Petrov <oleksandr.petrov@gmail.com>
Co-authored-by: Sam Tunnicliffe <samt@apache.org>