[CEP-21] Bulk update of unit and dtests
Updates the existing unit and and dtests to work with TCM. In the vast
majority of cases, this just means changes to initialization or to
slightly updated method signatures.

In CEP-21 generally, the intention has been not to modify existing
public interfaces at all and to limit any changes to code on the
boundaries of internal subsystems. In addition, care has been taken to
only make minimal modifications to existing tests, and to preserve their
invariants. So although this commit is fairly large in terms of
number of files, it's changes should be semantically quite light.

Co-authored-by: Marcus Eriksson <marcuse@apache.org>
Co-authored-by: Alex Petrov <oleksandr.petrov@gmail.com>
Co-authored-by: Sam Tunnicliffe <samt@apache.org>